Fix assessment import rename references
